Add feature flag wasmedge and rename shim with general keyword wasmwasi
