+++ Makefile
@@ -38,9 +38,9 @@ clean:
 
 .PHONY: install
 install:
-	$(INSTALL) target/$(TARGET)/containerd-shim-wasmedge-v1 $(PREFIX)/bin
-	$(INSTALL) target/$(TARGET)/containerd-shim-wasmedged-v1 $(PREFIX)/bin
-	$(INSTALL) target/$(TARGET)/containerd-wasmedged $(PREFIX)/bin
+	$(INSTALL) target/$(TARGET)/containerd-shim-wasmwasi-v1 $(PREFIX)/bin
+	$(INSTALL) target/$(TARGET)/containerd-shim-wasmwasid-v1 $(PREFIX)/bin
+	$(INSTALL) target/$(TARGET)/containerd-wasmwasid $(PREFIX)/bin
 
 # TODO: build this manually instead of requiring buildx
 test/out/img.tar: test/image/Dockerfile test/image/src/main.rs test/image/Cargo.toml test/image/Cargo.lock
