--- src/bin/containerd-shim-wasmedge-v1/main.rs
@@ -1,7 +1,0 @@
-use containerd_shim as shim;
-use containerd_shim_wasm::sandbox::ShimCli;
-use runwasi::instance::Wasi as WasiInstance;
-
-fn main() {
-    shim::run::<ShimCli<WasiInstance, wasmedge_sdk::Vm>>("io.containerd.wasmedge.v1", None);
-}
