+++ src/bin/containerd-shim-wasmwasid-v1/main.rs
@@ -2,5 +2,5 @@ use containerd_shim as shim;
 use containerd_shim_wasm::sandbox::manager::Shim;
 
 fn main() {
-    shim::run::<Shim>("containerd-shim-wasmedged-v1", None)
+    shim::run::<Shim>("containerd-shim-wasmwasid-v1", None)
 }
