+++ test/image/Cargo.toml
@@ -1,6 +1,6 @@
 
 [package]
-name = "containerd-shim-wasmedge-demo"
+name = "containerd-shim-wasmwasi-demo"
 version = "0.1.0"
 
 [workspace]
